pyproject.toml
[project]
name = "imgcapt"
version = "1.0.0"
description = "Professional FLUX LoRA training dataset preparation with AI-powered captioning"
authors = [
    {name = "IMGCAPT Contributors", email = ""}
]
readme = "README.md"
license = {file = "LICENSE"}
keywords = ["flux", "lora", "training", "captions", "ai", "ollama", "dataset"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "pillow>=10.2.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.2.0",
    "requests>=2.31.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/imgcapt"
Repository = "https://github.com/yourusername/imgcapt"
Issues = "https://github.com/yourusername/imgcapt/issues"
Documentation = "https://github.com/yourusername/imgcapt#readme"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]
